Trying to fix, or get more information out of, windows clang-cl
CI appears to be failing on api_listener_impl.cc, even though I didn't
touch it, so I'm pulling out the bits that it needs from the code I
did touch to isolate the changes.
