+++ source/server/filter_chain_manager_impl.cc
@@ -9,6 +9,7 @@
 #include "source/common/network/socket_interface.h"
 #include "source/common/protobuf/utility.h"
 #include "source/server/configuration_impl.h"
+#include "source/server/factory_context_impl.h"
 
 #include "absl/container/node_hash_map.h"
 #include "absl/strings/match.h"
@@ -746,50 +747,5 @@ FactoryContextImpl::FactoryContextImpl(Server::Instance& server,
     : server_(server), config_(config), drain_decision_(drain_decision),
       global_scope_(global_scope), listener_scope_(listener_scope) {}
 
-AccessLog::AccessLogManager& FactoryContextImpl::accessLogManager() {
-  return server_.accessLogManager();
-}
-Upstream::ClusterManager& FactoryContextImpl::clusterManager() { return server_.clusterManager(); }
-Event::Dispatcher& FactoryContextImpl::dispatcher() { return server_.dispatcher(); }
-const Server::Options& FactoryContextImpl::options() { return server_.options(); }
-Grpc::Context& FactoryContextImpl::grpcContext() { return server_.grpcContext(); }
-Router::Context& FactoryContextImpl::routerContext() { return server_.routerContext(); }
-bool FactoryContextImpl::healthCheckFailed() { return server_.healthCheckFailed(); }
-Http::Context& FactoryContextImpl::httpContext() { return server_.httpContext(); }
-Init::Manager& FactoryContextImpl::initManager() { return server_.initManager(); }
-const LocalInfo::LocalInfo& FactoryContextImpl::localInfo() const { return server_.localInfo(); }
-Envoy::Runtime::Loader& FactoryContextImpl::runtime() { return server_.runtime(); }
-Stats::Scope& FactoryContextImpl::scope() { return global_scope_; }
-Singleton::Manager& FactoryContextImpl::singletonManager() { return server_.singletonManager(); }
-OverloadManager& FactoryContextImpl::overloadManager() { return server_.overloadManager(); }
-ThreadLocal::SlotAllocator& FactoryContextImpl::threadLocal() { return server_.threadLocal(); }
-Admin& FactoryContextImpl::admin() { return server_.admin(); }
-TimeSource& FactoryContextImpl::timeSource() { return server_.timeSource(); }
-ProtobufMessage::ValidationContext& FactoryContextImpl::messageValidationContext() {
-  return server_.messageValidationContext();
-}
-ProtobufMessage::ValidationVisitor& FactoryContextImpl::messageValidationVisitor() {
-  return server_.messageValidationContext().staticValidationVisitor();
-}
-Api::Api& FactoryContextImpl::api() { return server_.api(); }
-ServerLifecycleNotifier& FactoryContextImpl::lifecycleNotifier() {
-  return server_.lifecycleNotifier();
-}
-ProcessContextOptRef FactoryContextImpl::processContext() { return server_.processContext(); }
-Configuration::ServerFactoryContext& FactoryContextImpl::getServerFactoryContext() const {
-  return server_.serverFactoryContext();
-}
-Configuration::TransportSocketFactoryContext&
-FactoryContextImpl::getTransportSocketFactoryContext() const {
-  return server_.transportSocketFactoryContext();
-}
-const envoy::config::core::v3::Metadata& FactoryContextImpl::listenerMetadata() const {
-  return config_.metadata();
-}
-envoy::config::core::v3::TrafficDirection FactoryContextImpl::direction() const {
-  return config_.traffic_direction();
-}
-Network::DrainDecision& FactoryContextImpl::drainDecision() { return drain_decision_; }
-Stats::Scope& FactoryContextImpl::listenerScope() { return listener_scope_; }
 } // namespace Server
 } // namespace Envoy
